requirements.txt
requests==2.31.0
pandas==2.1.4
google-cloud-bigquery==3.14.1
python-dotenv==1.0.0
pyyaml==6.0.1
entsoe-py==0.5.10
pyarrow==15.0.0